Refactor - Move SBPF program crate local tests into workspace global test directory (#1708)
* Splits SBPF program test crate `mem` into `dep`.

* Moves crate local tests into workspace global test directory.

* Removes non-dev dependencies from Cargo.toml
