+++ programs/sbf/rust/membuiltins/Cargo.toml
@@ -10,7 +10,7 @@ edition = { workspace = true }
 
 [dependencies]
 solana-program = { workspace = true }
-solana-sbf-rust-mem = { workspace = true, features = ["no-entrypoint"] }
+solana-sbf-rust-mem-dep = { workspace = true }
 
 [lib]
 crate-type = ["cdylib"]
