+++ programs/sbf/rust/membuiltins/src/lib.rs
@@ -6,7 +6,7 @@
 extern crate compiler_builtins;
 use {
     solana_program::{custom_heap_default, custom_panic_default, entrypoint::SUCCESS},
-    solana_sbf_rust_mem::{run_mem_tests, MemOps},
+    solana_sbf_rust_mem_dep::{run_mem_tests, MemOps},
 };
 
 #[no_mangle]
