+++ programs/sbf/rust/sanity/Cargo.toml
@@ -8,17 +8,8 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[features]
-test-bpf = []
-dummy-for-ci-check = ["test-bpf"]
-
 [dependencies]
 solana-program = { workspace = true }
-
-[dev-dependencies]
-solana-program-runtime = { workspace = true }
-solana-program-test = { workspace = true }
-solana-sdk = { workspace = true }
 
 [lib]
 crate-type = ["cdylib", "lib"]
