+++ programs/sbf/rust/simulation/Cargo.toml
@@ -8,17 +8,8 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
-[features]
-test-bpf = []
-
 [dependencies]
 solana-program = { workspace = true }
-
-[dev-dependencies]
-agave-validator = { workspace = true }
-solana-logger = { workspace = true }
-solana-program-test = { workspace = true }
-solana-sdk = { workspace = true }
 
 [lib]
 crate-type = ["cdylib", "lib"]
