+++ programs/sbf/tests/sanity.rs
